+++ src/Views/Dashboard/Dashboard.aspx.cs
@@ -11,6 +11,7 @@ namespace lovis.Views.Dashboard
     {
         protected void Page_Load(object sender, EventArgs e)
         {
+            
 
         }
     }