+++ src/Views/Register/Register.aspx.designer.cs
@@ -7,10 +7,10 @@
 // </generado automáticamente>
 //------------------------------------------------------------------------------
 
-namespace lovis.Views.User {
+namespace lovis.Views.Register {
     
     
-    public partial class User {
+    public partial class Register {
         
         /// <summary>
         /// Control form1.
@@ -22,60 +22,66 @@ namespace lovis.Views.User {
         protected global::System.Web.UI.HtmlControls.HtmlForm form1;
         
         /// <summary>
-        /// Control user_username.
+        /// Control reg_email.
         /// </summary>
         /// <remarks>
         /// Campo generado automáticamente.
         /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
         /// </remarks>
-        protected global::System.Web.UI.WebControls.TextBox user_username;
+        protected global::System.Web.UI.WebControls.TextBox reg_email;
         
         /// <summary>
-        /// Control user_password.
+        /// Control reg_password.
         /// </summary>
         /// <remarks>
         /// Campo generado automáticamente.
         /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
         /// </remarks>
-        protected global::System.Web.UI.WebControls.TextBox user_password;
+        protected global::System.Web.UI.WebControls.TextBox reg_password;
         
         /// <summary>
-        /// Control user_repeatpassword.
+        /// Control reg_repeatpassword.
         /// </summary>
         /// <remarks>
         /// Campo generado automáticamente.
         /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
         /// </remarks>
-        protected global::System.Web.UI.WebControls.TextBox user_repeatpassword;
+        protected global::System.Web.UI.WebControls.TextBox reg_repeatpassword;
         
         /// <summary>
-        /// Control user_name.
+        /// Control reg_name.
         /// </summary>
         /// <remarks>
         /// Campo generado automáticamente.
         /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
         /// </remarks>
-        protected global::System.Web.UI.WebControls.TextBox user_name;
+        protected global::System.Web.UI.WebControls.TextBox reg_name;
         
         /// <summary>
-        /// Control user_surname.
+        /// Control reg_surname.
         /// </summary>
         /// <remarks>
         /// Campo generado automáticamente.
         /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
         /// </remarks>
-        protected global::System.Web.UI.WebControls.TextBox user_surname;
+        protected global::System.Web.UI.WebControls.TextBox reg_surname;
         
         /// <summary>
-        /// Propiedad Master.
+        /// Control reg_checkbox.
         /// </summary>
         /// <remarks>
-        /// Propiedad generada automáticamente.
+        /// Campo generado automáticamente.
+        /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
         /// </remarks>
-        public new lovis.Views.Shared.Master Master {
-            get {
-                return ((lovis.Views.Shared.Master)(base.Master));
-            }
-        }
+        protected global::System.Web.UI.HtmlControls.HtmlInputCheckBox reg_checkbox;
+        
+        /// <summary>
+        /// Control reg_register.
+        /// </summary>
+        /// <remarks>
+        /// Campo generado automáticamente.
+        /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.Button reg_register;
     }
 }